Split picker and preview into separate components
